Rearranged files and added new function
