--- heroku_n_quagga/public/index.html
@@ -1,100 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
-
-    <title>index</title>
-    <meta name="description" content="" />
-    <meta name="author" content="Christoph Oberhofer" />
-
-    <meta name="viewport" content="width=device-width; initial-scale=1.0; user-scalable=no" />
-    <link rel="stylesheet" type="text/css" href="css/styles.css" />
-  </head>
-
-  <body>
-      <header>
-          <div class="headline">
-              <h1>QuaggaJS</h1>
-              <h2>An advanced barcode-scanner written in JavaScript</h2>
-          </div>
-      </header>
-    <section id="container" class="container">
-        <h3>The user's camera</h3>
-        <p>
-            Find the barcode to an item and scan at least 5 inches away.
-        </p>
-        <div class="controls">
-            <fieldset class="input-group">
-                <button class="stop">Stop</button>
-            </fieldset>
-            <fieldset class="reader-config-group">
-                <label>
-                    <span>Barcode-Type</span>
-                    <select name="decoder_readers">
-                        <option value="code_128" selected="selected">Code 128</option>
-                    </select>
-                </label>
-                <label>
-                    <span>Resolution (width)</span>
-                    <select name="input-stream_constraints">
-                        <option value="320x240">320px</option>
-                        <option selected="selected" value="640x480">640px</option>
-                        <option value="800x600">800px</option>
-                        <option value="1280x720">1280px</option>
-                        <option value="1600x960">1600px</option>
-                        <option value="1920x1080">1920px</option>
-                    </select>
-                </label>
-                <label>
-                    <span>Patch-Size</span>
-                    <select name="locator_patch-size">
-                        <option value="x-small">x-small</option>
-                        <option value="small">small</option>
-                        <option selected="selected" value="medium">medium</option>
-                        <option value="large">large</option>
-                        <option value="x-large">x-large</option>
-                    </select>
-                </label>
-                <label>
-                    <span>Half-Sample</span>
-                    <input type="checkbox" checked="checked" name="locator_half-sample" />
-                </label>
-                <label>
-                    <span>Workers</span>
-                    <select name="numOfWorkers">
-                        <option value="0">0</option>
-                        <option value="1">1</option>
-                        <option value="2">2</option>
-                        <option selected="selected" value="4">4</option>
-                        <option value="8">8</option>
-                    </select>
-                </label>
-                <label>
-                    <span>Camera</span>
-                    <select name="input-stream_constraints" id="deviceSelection">
-                    </select>
-                </label>
-                <label style="display: none">
-                    <span>Zoom</span>
-                    <select name="settings_zoom"></select>
-                </label>
-                <label style="display: none">
-                    <span>Torch</span>
-                    <input type="checkbox" name="settings_torch" />
-                </label>
-            </fieldset>
-        </div>
-      <div id="result_strip">
-        <ul class="thumbnails"></ul>
-        <ul class="collector"></ul>
-      </div>
-      <div id="interactive" class="viewport"></div>
-    </section>
-    <script
-        src="https://code.jquery.com/jquery-3.3.1.min.js" integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8=" crossorigin="anonymous"></script>
-    <script src="//webrtc.github.io/adapter/adapter-latest.js" type="text/javascript"></script>
-    <script src="app.js" type="text/javascript"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/quagga/0.12.1/quagga.js"></script>
-  </body>
-</html>
